feat: token usage speed/count
